Backports
Backported - GS: Fix zequal handling in CDrawScanline
https://github.com/PCSX2/pcsx2/commit/650dc8bc6a768655d10b62935f4877b6add5d1b8

Backported - GS: Add zequal to GSScanlineSelector to_string
https://github.com/PCSX2/pcsx2/commit/4c298ca484c125ff3282aad7b822edb03f18e438

Backported - Config: Fix frame rate being parsed ever-so-slightly off
https://github.com/PCSX2/pcsx2/commit/22e9aea827987384ed795e2f2deca1e2abd0af0d

Backported - VMManager: Only trigger hotkeys with valid VM
https://github.com/PCSX2/pcsx2/commit/6fca0dac0ea1326ca4c524ccdf9776f3a2adab02

Backported - PAD: Fix copying input sources
https://github.com/PCSX2/pcsx2/commit/3d6f4629e9210d883f2ef158937df5e9d0dc9029

Backported - Host: Add GetResourceFileTimestamp()
https://github.com/PCSX2/pcsx2/commit/a228582984988b9e13a87d5463ed83489c1039d8

Backported - HostDisplay: Make GL/D3D timestamp queries slightly less rubbish
https://github.com/PCSX2/pcsx2/commit/6e0ac7fa75fbd75471701a5e76ee76a6c2eee252

Backported - VMManager: Add warning for unsafe settings on boot
https://github.com/PCSX2/pcsx2/commit/43d6ddd7cf2334085bdb878b1a861cb3cd4fb4fd

Backported - GzippedFileReader: Don't double open when creating index
https://github.com/PCSX2/pcsx2/commit/e8704adc6754a0b7659c995eef7c5b2af70c7c27

Backported - LRUCache: Fix Evict() removing too many items
https://github.com/PCSX2/pcsx2/commit/e0cb16592797989d441a0be3f2e9708853f3d0be

Backported - iR5900Analysis: Always set COP2 status flag bit for VSQRT/VRSQRT/VDIV
https://github.com/PCSX2/pcsx2/commit/5a0c8c9f324d96b97d25796d3dcd74af2af5d06b

Backported - iR5900: Skip reloading COP2 flags register when it's not used
https://github.com/PCSX2/pcsx2/commit/bc3729c930190e4732af9ffc7eb9b7cc4820b6c5

Backported - UI: Fullscreen UI typo fix
https://github.com/PCSX2/pcsx2/commit/72c78fc163469804c18b47d416aecb0aedf3e8af

+++ common/SettingsWrapper.h
@@ -28,7 +28,7 @@ public:
 	virtual void Entry(const char* section, const char* var, int& value, const int defvalue = 0) = 0;
 	virtual void Entry(const char* section, const char* var, uint& value, const uint defvalue = 0) = 0;
 	virtual void Entry(const char* section, const char* var, bool& value, const bool defvalue = false) = 0;
-	virtual void Entry(const char* section, const char* var, double& value, const double defvalue = 0.0) = 0;
+	virtual void Entry(const char* section, const char* var, float& value, const float defvalue = 0.0) = 0;
 	virtual void Entry(const char* section, const char* var, std::string& value, const std::string& default_value = std::string()) = 0;
 
 	// This special form of Entry is provided for bitfields, which cannot be passed by reference.
@@ -64,7 +64,7 @@ public:
 	void Entry(const char* section, const char* var, int& value, const int defvalue = 0) override;
 	void Entry(const char* section, const char* var, uint& value, const uint defvalue = 0) override;
 	void Entry(const char* section, const char* var, bool& value, const bool defvalue = false) override;
-	void Entry(const char* section, const char* var, double& value, const double defvalue = 0.0) override;
+	void Entry(const char* section, const char* var, float& value, const float defvalue = 0.0) override;
 
 	void Entry(const char* section, const char* var, std::string& value, const std::string& default_value = std::string()) override;
 	bool EntryBitBool(const char* section, const char* var, bool value, const bool defvalue = false) override;
@@ -85,7 +85,7 @@ public:
 	void Entry(const char* section, const char* var, int& value, const int defvalue = 0) override;
 	void Entry(const char* section, const char* var, uint& value, const uint defvalue = 0) override;
 	void Entry(const char* section, const char* var, bool& value, const bool defvalue = false) override;
-	void Entry(const char* section, const char* var, double& value, const double defvalue = 0.0) override;
+	void Entry(const char* section, const char* var, float& value, const float defvalue = 0.0) override;
 
 	void Entry(const char* section, const char* var, std::string& value, const std::string& default_value = std::string()) override;
 	bool EntryBitBool(const char* section, const char* var, bool value, const bool defvalue = false) override;
